requirements.txt
fastapi
uvicorn
openai
langchain>=1.2.0
langchain-openai>=1.1.0
langchain-community>=0.3.0
langchain-core>=1.2.0
langgraph>=1.1.0
chromadb>=0.6.0
pypdf>=5.0.0
openpyxl>=3.1.0
pandas>=2.2.0
python-dotenv>=1.0.0
streamlit>=1.45.0